% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GdClean.R
\name{GdClean}
\alias{GdClean}
\title{Removing Gd contamination in CyTOF data.}
\usage{
GdClean(fs, gdRatio = NULL, method = "1DNorm")
}
\arguments{
\item{fs}{The flowSet object of contaminated CyTOF files.}

\item{gdRatio}{The selected intensity ratios across Gd channels, default with intensity ratios of natural abundance}

\item{method}{The method used for calculating contamination coefficients for each cell, default with '1DNorm', available options are '1DNorm','2DNorm','Min','Mean','Median'}
}
\value{
The flowSet object of cleaned CyTOF files.
}
\description{
Removing Gd contamination in CyTOF data.
}
\examples{
library(GdClean)

fcs.dir <- "../fcsData/"
frames <- lapply(dir(fcs.dir, full.names = T), read.FCS, transformation = F, alter.names = T)
names(frames) <- basename(dir(fcs.dir))
fs <- as(frames, "flowSet")

PercentList <- seq(5, 100, 5)
GdRatios <- estimateGdRatio(ff[[1]], percentList)

gdRatio <- GdRatios["5\%", ]
method <- "1DNorm"
fs_Clean <- GdClean(fs, gdRatio = gdRatio, method = method)
}
