% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GdCorrHeatmap.R
\name{GdCorrHeatmap}
\alias{GdCorrHeatmap}
\title{Calculate the Pearson correlation coefficients across Gd channels in each CyTOF data in the flowSet object.}
\usage{
GdCorrHeatmap(fs)
}
\arguments{
\item{fs}{The input flowSet object}
}
\value{
The output of pheatmap object
}
\description{
Calculate the Pearson correlation coefficients across Gd channels in each CyTOF data in the flowSet object.
}
\examples{
library(GdClean)
p1 <- GdCorrHeatmap(fs)
p1
}
