% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GdRatioPlot.R
\name{GdRatioPlot}
\alias{GdRatioPlot}
\title{Compare the calculated Gd ratios with the Gd ratios of natural abundance}
\usage{
GdRatioPlot(GdRatios)
}
\arguments{
\item{GdRatios}{The calculated Gd Ratios from the 'estimateGdRatio'
'}
}
\value{
The ggplot object of the compared result.
}
\description{
Compare the calculated Gd ratios with the Gd ratios of natural abundance
}
\examples{
library(GdClean)
p1 <- GdRatioPlot(GdRatios)
p1
}
