% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGdChannels.R
\name{getGdChannels}
\alias{getGdChannels}
\title{Return the information of used Gadolinium channels in CyTOF data}
\usage{
getGdChannels(ff)
}
\arguments{
\item{ff}{flowFrame object}
}
\value{
dataFrame of Gd channel ID and related metal Isotope
}
\description{
Return the information of used Gadolinium channels in CyTOF data
}
\examples{
library(GdClean)
gdInfo <- getGdChannels(ff)
}
