% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMassChannels.R
\name{getMassChannels}
\alias{getMassChannels}
\title{Return the information of channels with Gd isotope mass in CyTOF data.}
\usage{
getMassChannels(ff)
}
\arguments{
\item{ff}{flowFrame object}
}
\value{
dataFrame of channel information within the mass range of Gd isotopes.
}
\description{
Return the information of channels with Gd isotope mass in CyTOF data.
}
\examples{
library(GdClean)
massInfo <- getMassChannels(ff)
}
