% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGdRatio.R
\name{estimateGdRatio}
\alias{estimateGdRatio}
\title{Estimating the intensity ratios across channels within mass range of Gd, the first 155 channel was selected as the reference channel.}
\usage{
estimateGdRatio(ff, PercentList)
}
\arguments{
\item{ff}{The flowframe object of Gd contaminated CyTOF data.}

\item{PercentList}{The top-n percent of cells selected for estimating Gd ratios, sorted by mean expression of Gd channels.}
}
\value{
The estimated intensity ratios across channels, the 155 channel was selected as the reference channel.
}
\description{
Estimating the intensity ratios across channels within mass range of Gd, the first 155 channel was selected as the reference channel.
}
\examples{
library(GdClean)

PercentList <- seq(5, 100, 5)
GdRatios <- estimateGdRatio(ff, percentList)
}
