% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GdRatioPlot.R
\name{GdRatioPlot}
\alias{GdRatioPlot}
\title{Comparing the estimated Gd ratios with the of natural Gd abundance ratios}
\usage{
GdRatioPlot(GdRatios)
}
\arguments{
\item{GdRatios}{The calculated Gd Ratios from the 'estimateGdRatio'
'}
}
\value{
The ggplot object of the compared result.
}
\description{
Comparing the estimated Gd ratios with the of natural Gd abundance ratios
}
\examples{
library(GdClean)
p1 <- GdRatioPlot(GdRatios)
p1
}
